+++ tests/visual/text.spec.js
@@ -43,4 +43,11 @@ describe('text', function() {
       doc.fillColor('#000').list(['One', 'Two', 'Three'], 100, 150);
     });
   });
+
+  test('list with line breaks in items', function() {
+    return runDocTest(function(doc) {
+      doc.font('tests/fonts/Roboto-Regular.ttf');
+      doc.list(['Foo\nBar', 'Foo\rBar', 'Foo\r\nBar'], [100, 150]);
+    })
+  })
 });